add model runner back to util
