+++ ucb_prefect_tools/util.py
@@ -26,6 +26,10 @@ from prefect.client.orchestration import get_client
 import git
 import pytz
 
+import rpy2.robjects as ro
+from rpy2.robjects.packages import importr
+from rpy2.robjects import pandas2ri
+
 # Overrideable settings related to deployments
 DOCKER_REGISTRY = "oit-data-services-docker-local.artifactory.colorado.edu"
 LOCAL_FLOW_FOLDER = "flows"
@@ -369,5 +373,20 @@ def sizeof_fmt(num: int) -> str:
         num /= 1024.0
     return f"{num:.1f} PB"
 
+@task
+def run_model(data: pd.DataFrame, model_path:str) -> pd.DataFrame:
+    """Reads an R model from an RDS file and runs its associated predict method on a dataframe, returning the predictions"""
+    importr("tidyverse")
+    importr("tidymodels")
+    importr("magrittr")
+    importr("arrow")
+    importr("iml")
+    importr("data.table")
+    importr("caret")
 
-
+    with (ro.default_converter + pandas2ri.converter).context():
+        r_dataframe = ro.conversion.get_conversion().py2rpy(data)
+        model = ro.r.readRDS(model_path)
+        preds = ro.r.predict(model, r_dataframe)
+        pred_dataframe = ro.conversion.get_conversion().rpy2py(preds)
+    return pred_dataframe
